Updated Text-to-speech
Before the program would use system() to call flite to make text-to-speech, this would pause the execution of the program until the synthesis was complete which caused the program to appear unresponsive for seconds at a time. The new function places the text to be said into a pipe which is then read in by a second program which is responsible for calling espeak.